[fix] transaction 적용 안되는 문제 해결
