+++ src/main/java/org/example/pay/auth/service/AuthService.java
@@ -9,8 +9,6 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 
 @Service
@@ -20,7 +18,6 @@ public class AuthService {
 
 	private final MemberRepository memberRepository;
 	private final BCryptPasswordEncoder bCryptPasswordEncoder;
-	private final SecurityContextService securityContextService;
 
 	@Transactional
 	public Member join(JoinRequestDto joinRequestDto) {
@@ -37,31 +34,5 @@ public class AuthService {
 		memberRepository.save(member);
 
 		return member;
-	}
-
-	@Transactional
-	public String tempLogin(LoginRequestDto tempLoginDto, HttpServletRequest request) {
-
-		Member member = memberRepository.findByEmail(tempLoginDto.email())
-			.orElseGet(() -> {
-				JoinRequestDto joinRequestDto = new JoinRequestDto(tempLoginDto.email(), "임시유저", tempLoginDto.password());
-				return join(joinRequestDto);
-			});
-
-		if (!bCryptPasswordEncoder.matches(tempLoginDto.password(), member.getPassword())) {
-			throw new IllegalArgumentException("비밀번호가 일치하지 않습니다.");
-		}
-
-		HttpSession session = createSession(request, member);
-
-		securityContextService.saveSecurityContext(session, member);
-
-		return session.getId();
-	}
-
-	private HttpSession createSession(HttpServletRequest request, Member member) {
-		HttpSession session = request.getSession(true);
-		session.setAttribute("loginMemberId", member.getId());
-		return session;
 	}
 }
